refactor(adaptive-options): move type and file
