--- src/shared/types/index.ts
@@ -1,9 +1,0 @@
-export interface AdaptiveOptions {
-  XXL: number
-  XL: number
-  LG: number
-  MD: number
-  SM: number
-  XS: number
-  XXS: number
-}
